+++ django_ninja_admin/admins/auth.py
@@ -1,6 +1,11 @@
 from django_ninja_admin.admins.model import ModelAdmin
 
 
+def _existing_model_fields(model, field_names):
+    model_fields = {field.name for field in (*model._meta.fields, *model._meta.many_to_many)}
+    return tuple(field for field in field_names if field in model_fields)
+
+
 class AuthUserAdmin(ModelAdmin):
     list_display = ("__str__", "is_active", "is_staff")
     search_fields = ("username", "email", "first_name", "last_name")
@@ -8,8 +13,12 @@ class AuthUserAdmin(ModelAdmin):
 
     def get_exclude(self, request, obj=None):
         excluded = list(super().get_exclude(request, obj) or [])
-        model_fields = {field.name for field in self.model._meta.get_fields()}
-        excluded.extend(field for field in self.sensitive_fields if field in model_fields)
+        excluded.extend(_existing_model_fields(self.model, self.sensitive_fields))
+        return tuple(dict.fromkeys(excluded))
+
+    def get_output_exclude(self, request=None):
+        excluded = list(super().get_output_exclude(request) or [])
+        excluded.extend(_existing_model_fields(self.model, self.sensitive_fields))
         return tuple(dict.fromkeys(excluded))
 
 
@@ -20,6 +29,10 @@ class AuthGroupAdmin(ModelAdmin):
 
     def get_exclude(self, request, obj=None):
         excluded = list(super().get_exclude(request, obj) or [])
-        model_fields = {field.name for field in self.model._meta.get_fields()}
-        excluded.extend(field for field in self.sensitive_fields if field in model_fields)
+        excluded.extend(_existing_model_fields(self.model, self.sensitive_fields))
+        return tuple(dict.fromkeys(excluded))
+
+    def get_output_exclude(self, request=None):
+        excluded = list(super().get_output_exclude(request) or [])
+        excluded.extend(_existing_model_fields(self.model, self.sensitive_fields))
         return tuple(dict.fromkeys(excluded))